Major update

+++ nlmod/cache.py
@@ -43,14 +43,16 @@ def clear_cache(cachedir):
 
             # remove pklz file
             os.remove(os.path.join(cachedir, fname))
+            logger.info(f"removed {fname}")
 
-            # remove netcdf file (make sure cached netcdf is closed)
+            # remove netcdf file
             fpath_nc = os.path.join(cachedir, fname_nc)
-            cached_ds = xr.open_dataset(fpath_nc)
-            cached_ds.close()
-
-            os.remove(fpath_nc)
-            logger.info(f"removing {fname} and {fname_nc}")
+            if os.path.exists(fname_nc):
+                # make sure cached netcdf is closed
+                cached_ds = xr.open_dataset(fpath_nc)
+                cached_ds.close()
+                os.remove(fpath_nc)
+                logger.info(f"removed {fname_nc}")
 
 
 def cache_netcdf(func):
@@ -89,6 +91,7 @@ def cache_netcdf(func):
     @functools.wraps(func)
     def decorator(*args, cachedir=None, cachename=None, **kwargs):
 
+        # 1 check if cachedir and name are provided
         if cachedir is None or cachename is None:
             return func(*args, **kwargs)
 
@@ -96,9 +99,7 @@ def cache_netcdf(func):
             cachename += ".nc"
 
         fname_cache = os.path.join(cachedir, cachename)  # netcdf file
-        fname_pickle_cache = fname_cache.replace(
-            ".nc", ".pklz"
-        )  # pickle with function arguments
+        fname_pickle_cache = fname_cache.replace(".nc", ".pklz")
 
         # create dictionary with function arguments
         func_args_dic = {f"arg{i}": args[i] for i in range(len(args))}
@@ -281,7 +282,9 @@ def _check_ds(ds, ds2):
     """
     for coord in ds2.coords:
         if coord in ds.coords:
-            if not ds2[coord].equals(ds[coord]):
+            try:
+                xr.testing.assert_identical(ds[coord], ds2[coord])
+            except AssertionError:
                 logger.info(
                     f"coordinate {coord} has different values in cached dataset, not using cache"
                 )
